+++ gata-frontend/app/components/ui/dropdown-menu.tsx
@@ -165,15 +165,6 @@ export {
    DropdownMenuTrigger,
    DropdownMenuContent,
    DropdownMenuItem,
-   DropdownMenuCheckboxItem,
-   DropdownMenuRadioItem,
    DropdownMenuLabel,
    DropdownMenuSeparator,
-   DropdownMenuShortcut,
-   DropdownMenuGroup,
-   DropdownMenuPortal,
-   DropdownMenuSub,
-   DropdownMenuSubContent,
-   DropdownMenuSubTrigger,
-   DropdownMenuRadioGroup,
 };
