+++ gata-frontend/package.json
@@ -18,10 +18,7 @@
    "dependencies": {
       "@radix-ui/react-avatar": "^1.0.4",
       "@radix-ui/react-dropdown-menu": "^2.0.6",
-      "@radix-ui/react-label": "^2.0.2",
-      "@radix-ui/react-progress": "^1.0.3",
       "@radix-ui/react-slot": "^1.0.2",
-      "@radix-ui/react-toggle": "^1.0.3",
       "@radix-ui/react-toggle-group": "^1.0.4",
       "@radix-ui/react-tooltip": "^1.0.7",
       "@remix-run/cloudflare": "^2.8.1",
@@ -29,8 +26,6 @@
       "@remix-run/react": "2.8.1",
       "class-variance-authority": "^0.7.0",
       "clsx": "^2.1.0",
-      "framer-motion": "11.0.25",
-      "isbot": "^4",
       "lucide-react": "^0.363.0",
       "miniflare": "3.20240404.0",
       "react": "18.2.0",
@@ -42,8 +37,7 @@
       "slate-history": "0.100.0",
       "slate-react": "0.102.0",
       "tailwind-merge": "^2.2.2",
-      "tailwindcss-animate": "^1.0.7",
-      "zod": "^3.22.5"
+      "tailwindcss-animate": "^1.0.7"
    },
    "devDependencies": {
       "@cloudflare/workers-types": "^4.20240320.1",
